Add files via upload
Each file is labelled "fig_5" followed by either "case_1" or "case_2", meaning l^2 is either positive or negative respectively, followed by the value of l^2 as listed in the legend of figure 5. Each file contains the variable "l_sq", denoting the value of l^2, followed by the arrays "C_L" and "C_R", referring to the left and right principal curves respectively that are plotted in figure 5. They each contain two rows of points, the first row corresponds to the x-coordinates and the second to the y-coordinates.