Add files via upload
Each file is labelled "fig_6_eps" followed by the value of epsilon corresponding to those listed in figure 6. Each file contains the value of epsilon, stored in "eps", followed by the first 20 pole locations on the left and right singularity strings, stored in the arrays "z_sing_L" and "z_sing_R" respectively. These arrays each have two rows, the first row corresponds to the x-coordinates and the second to the y-coordinates. The left and right principal curves are also in these files, stored in the arrays "C_L" and "C_R" respectively. Likewise they each have two rows, with the x-coordinates stored in the first row and the y-coordinates in the second. Finally the file "fig_6_error" contains the first 20 singularities for the complementary error function taken from the DLMF that are used to validate the accuracy of our approximations when a = 0.5, stored in the array "z_erfc" -- see the main paper body for a full explanation on the use of erfc(z) and the calculation of the error. The arrays "z_L_O" and "z_1_X" contain the leading order and first order approximations to the pole locations. All three of these arrays have two rows, with the x-coordinates stored in the first row and the y-coordinates in the second row.